refactor: move announcement logic to AnnounceBar
