+++ src/components/AnnouncementBar/styled.ts
