+++ src/components/AnnouncementBar/types.ts
